+++ examples/01_single_orbit_setup.py
@@ -3,7 +3,7 @@ import sys
 
 import agama
 
-from ocd_gd._logging_config import print_banner, print_kv_table, setup_logging
+from ocd_gd import print_banner, print_kv_table, setup_logging
 from ocd_gd.orbit_detector import OrbitChaosDetector